# speed up `get_args()` helper