--- a/R/translate.R
+++ b/R/translate.R
@@ -127,9 +127,12 @@ get_model_spec <- function(model, mode, engine) {
 
 get_args <- function(model, engine) {
   m_env <- get_model_env()
-  rlang::env_get(m_env, paste0(model, "_args")) %>%
-    dplyr::filter(engine == !!engine) %>%
-    dplyr::select(-engine)
+
+  args <- rlang::env_get(m_env, paste0(model, "_args"))
+  args <- vctrs::vec_slice(args, args$engine == engine)
+  args$engine <- NULL
+
+  args
 }
 
 # to replace harmonize